Refactor polygon drawing workflow and adjust labels
