+++ lib/features/geofence/presentation/pages/geofence_map_page.dart
@@ -9,6 +9,7 @@ import 'package:balumohol/features/geofence/models/custom_place.dart';
 import 'package:balumohol/features/geofence/models/location_history_entry.dart';
 import 'package:balumohol/features/geofence/models/polygon_feature.dart';
 import 'package:balumohol/features/geofence/models/user_polygon.dart';
+import 'package:balumohol/features/geofence/presentation/pages/draw_polygon_page.dart';
 import 'package:balumohol/features/geofence/presentation/widgets/current_location_indicator.dart';
 import 'package:balumohol/features/geofence/presentation/widgets/custom_place_marker.dart';
 import 'package:balumohol/features/geofence/presentation/widgets/place_details_sheet.dart';
@@ -81,31 +82,6 @@ const List<_BaseLayerOption> _baseLayerOptions = <_BaseLayerOption>[
   ),
 ];
 
-class _PolygonColorOption {
-  const _PolygonColorOption(this.label, this.color);
-
-  final String label;
-  final Color color;
-}
-
-const List<_PolygonColorOption> _polygonColorOptions = <_PolygonColorOption>[
-  _PolygonColorOption('নীল', Color(0xFF1976D2)),
-  _PolygonColorOption('সবুজ', Color(0xFF2E7D32)),
-  _PolygonColorOption('কমলা', Color(0xFFEF6C00)),
-  _PolygonColorOption('বেগুনি', Color(0xFF6A1B9A)),
-  _PolygonColorOption('লাল', Color(0xFFC62828)),
-];
-
-class _NewPolygonDetails {
-  const _NewPolygonDetails({
-    required this.name,
-    required this.color,
-  });
-
-  final String name;
-  final Color color;
-}
-
 class _GeofenceMapPageState extends State<GeofenceMapPage> {
   static const double _customPlaceMarkerZoomThreshold = 14;
   static const double _customPlaceMarkerBaseWidth = 160;
@@ -116,8 +92,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
   CustomPlace? _selectedCustomPlace;
   bool _statusPanelCollapsed = true;
   _MapLayerType _selectedLayerType = _MapLayerType.hybrid;
-  bool _isDrawingPolygon = false;
-  final List<LatLng> _draftPolygonPoints = <LatLng>[];
 
   bool get _showCustomPlaceMarkers =>
       _currentZoom >= _customPlaceMarkerZoomThreshold;
@@ -154,10 +128,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
         : <Marker>[];
     final pathPoints = controller.trackingPath;
     final polygonLabelMarkers = _buildPolygonLabels(controller);
-    final draftMarkers =
-        _isDrawingPolygon ? _buildDraftPolygonMarkers() : const <Marker>[];
-    final draftPoints = List<LatLng>.from(_draftPolygonPoints);
-    final bool isDrawingPolygon = _isDrawingPolygon;
     final accuracyValue = controller.currentAccuracy;
     final accuracyText = accuracyValue != null
         ? formatMeters(accuracyValue, fractionDigits: 0)
@@ -228,12 +198,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
               initialCenter: controller.fallbackCenter,
               initialZoom: 15,
               onTap: (tapPosition, point) async {
-                if (_isDrawingPolygon) {
-                  setState(() {
-                    _draftPolygonPoints.add(point);
-                  });
-                  return;
-                }
                 final polygon = controller.polygonAt(point);
                 if (polygon != null) {
                   await _showPolygonDetails(controller, polygon);
@@ -261,33 +225,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
                 PolygonLayer(polygons: _buildPolygons(controller)),
               if (polygonLabelMarkers.isNotEmpty)
                 MarkerLayer(markers: polygonLabelMarkers),
-              if (isDrawingPolygon && draftPoints.length >= 3)
-                PolygonLayer(
-                  polygons: [
-                    Polygon(
-                      points: draftPoints,
-                      color: Colors.deepOrangeAccent.withOpacity(0.2),
-                      borderColor: Colors.deepOrangeAccent,
-                      borderStrokeWidth: 3,
-                      isFilled: true,
-                    ),
-                  ],
-                ),
-              if (isDrawingPolygon && draftPoints.length >= 2)
-                PolylineLayer(
-                  polylines: [
-                    Polyline(
-                      points: [
-                        ...draftPoints,
-                        if (draftPoints.length >= 3) draftPoints.first,
-                      ],
-                      strokeWidth: 3,
-                      color: Colors.deepOrangeAccent,
-                    ),
-                  ],
-                ),
-              if (isDrawingPolygon && draftMarkers.isNotEmpty)
-                MarkerLayer(markers: draftMarkers),
               if (pathPoints.length >= 2)
                 PolylineLayer(
                   polylines: [
@@ -305,10 +242,8 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
               if (customPlaceMarkers.isNotEmpty)
                 MarkerLayer(markers: customPlaceMarkers),
               if (currentMarker != null) MarkerLayer(markers: [currentMarker]),
-            ],
-          ),
-          if (isDrawingPolygon)
-            _buildPolygonDrawingOverlay(controller, draftPoints.length),
+              ],
+            ),
           SafeArea(
             child: Padding(
               padding: const EdgeInsets.all(12),
@@ -397,7 +332,13 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
   }
 
   List<Marker> _buildPolygonLabels(GeofenceMapController controller) {
-    final polygons = controller.polygons;
+    if (!controller.showBoundary) {
+      return const [];
+    }
+    final polygons = controller.polygons.where((polygon) {
+      final layerType = polygon.properties['layer_type'];
+      return layerType == 'boundary';
+    }).toList();
     if (polygons.isEmpty) {
       return const [];
     }
@@ -476,52 +417,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
     return markers;
   }
 
-  List<Marker> _buildDraftPolygonMarkers() {
-    return List<Marker>.generate(
-      _draftPolygonPoints.length,
-      (index) {
-        final point = _draftPolygonPoints[index];
-        return Marker(
-          point: point,
-          width: 36,
-          height: 36,
-          alignment: Alignment.center,
-          child: Builder(
-            builder: (context) {
-              final theme = Theme.of(context);
-              final textStyle = theme.textTheme.labelSmall?.copyWith(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  ) ??
-                  const TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                  );
-              return IgnorePointer(
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.deepOrangeAccent,
-                    shape: BoxShape.circle,
-                    border: Border.all(color: Colors.white, width: 2),
-                    boxShadow: const [
-                      BoxShadow(
-                        color: Colors.black26,
-                        blurRadius: 4,
-                        offset: Offset(0, 2),
-                      ),
-                    ],
-                  ),
-                  alignment: Alignment.center,
-                  child: Text('${index + 1}', style: textStyle),
-                ),
-              );
-            },
-          ),
-        );
-      },
-    );
-  }
-
   Marker? _buildCurrentLocationMarker(GeofenceMapController controller) {
     final location = controller.currentLocation;
     if (location == null) {
@@ -576,75 +471,32 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
         .toList();
   }
 
-  void _startPolygonDrawing(GeofenceMapController controller) {
-    if (_isDrawingPolygon) {
-      return;
-    }
+  Future<void> _startPolygonDrawing(GeofenceMapController controller) async {
     Scaffold.maybeOf(context)?.closeDrawer();
-    setState(() {
-      _isDrawingPolygon = true;
-      _draftPolygonPoints.clear();
-    });
     controller.highlightPolygon(null);
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(
-        content: Text('নতুন পলিগন আঁকার জন্য মানচিত্রে পয়েন্ট ট্যাপ করুন।'),
-        duration: Duration(seconds: 4),
+
+    final initialCenter = controller.currentLocation ??
+        controller.primaryCenter ??
+        controller.fallbackCenter;
+
+    final userPolygon = await Navigator.of(context).push<UserPolygon>(
+      MaterialPageRoute(
+        builder: (context) => DrawPolygonPage(
+          initialCenter: initialCenter,
+          tileUrlTemplate: _selectedBaseLayer.urlTemplate,
+          tileSubdomains: _selectedBaseLayer.subdomains,
+        ),
       ),
     );
-  }
 
-  void _cancelPolygonDrawing() {
-    if (!_isDrawingPolygon) {
+    if (!mounted || userPolygon == null) {
       return;
     }
-    setState(() {
-      _isDrawingPolygon = false;
-      _draftPolygonPoints.clear();
-    });
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('পলিগন আঁকা বাতিল করা হয়েছে।')),
-    );
-  }
-
-  void _undoPolygonPoint() {
-    if (_draftPolygonPoints.isEmpty) {
-      return;
-    }
-    setState(() {
-      _draftPolygonPoints.removeLast();
-    });
-  }
-
-  Future<void> _finishPolygonDrawing(GeofenceMapController controller) async {
-    if (_draftPolygonPoints.length < 3) {
-      return;
-    }
-    final details = await _promptForPolygonDetails();
-    if (!mounted || details == null) {
-      return;
-    }
-
-    final trimmedName = details.name.trim();
-    final displayName =
-        trimmedName.isEmpty ? 'কাস্টম পলিগন' : trimmedName;
-
-    final userPolygon = UserPolygon(
-      id: DateTime.now().millisecondsSinceEpoch.toString(),
-      name: displayName,
-      colorValue: details.color.value,
-      points: List<LatLng>.from(_draftPolygonPoints),
-    );
 
     await controller.addUserPolygon(userPolygon);
     if (!mounted) {
       return;
     }
-
-    setState(() {
-      _isDrawingPolygon = false;
-      _draftPolygonPoints.clear();
-    });
 
     final targetId = 'user_${userPolygon.id}';
     PolygonFeature? createdFeature;
@@ -655,6 +507,7 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
       }
     }
 
+    final displayName = userPolygon.name;
     if (createdFeature != null) {
       controller.highlightPolygon(createdFeature);
       final centroid = polygonCentroid(createdFeature);
@@ -665,157 +518,6 @@ class _GeofenceMapPageState extends State<GeofenceMapPage> {
 
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(content: Text('"$displayName" পলিগন সংরক্ষণ করা হয়েছে।')),
-    );
-  }
-
-  Future<_NewPolygonDetails?> _promptForPolygonDetails() async {
-    final nameController = TextEditingController();
-    Color selectedColor = _polygonColorOptions.first.color;
-    try {
-      return showDialog<_NewPolygonDetails>(
-        context: context,
-        builder: (context) {
-          return StatefulBuilder(
-            builder: (context, setState) {
-              final theme = Theme.of(context);
-              return AlertDialog(
-                title: const Text('পলিগনের বিস্তারিত'),
-                content: SingleChildScrollView(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      TextField(
-                        controller: nameController,
-                        decoration: const InputDecoration(
-                          labelText: 'নাম',
-                          hintText: 'পলিগনের নাম লিখুন',
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        'রং নির্বাচন করুন',
-                        style: theme.textTheme.titleSmall,
-                      ),
-                      const SizedBox(height: 12),
-                      Wrap(
-                        spacing: 16,
-                        runSpacing: 16,
-                        children: [
-                          for (final option in _polygonColorOptions)
-                            Material(
-                              color: Colors.transparent,
-                              child: InkWell(
-                                onTap: () => setState(
-                                  () => selectedColor = option.color,
-                                ),
-                                borderRadius: BorderRadius.circular(28),
-                                child: Padding(
-                                  padding: const EdgeInsets.symmetric(
-                                    horizontal: 4,
-                                  ),
-                                  child: _PolygonColorPreview(
-                                    color: option.color,
-                                    label: option.label,
-                                    selected: selectedColor == option.color,
-                                  ),
-                                ),
-                              ),
-                            ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-                actions: [
-                  TextButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    child: const Text('বাতিল'),
-                  ),
-                  FilledButton(
-                    onPressed: () {
-                      Navigator.of(context).pop(
-                        _NewPolygonDetails(
-                          name: nameController.text,
-                          color: selectedColor,
-                        ),
-                      );
-                    },
-                    child: const Text('সংরক্ষণ করুন'),
-                  ),
-                ],
-              );
-            },
-          );
-        },
-      );
-    } finally {
-      nameController.dispose();
-    }
-  }
-
-  Widget _buildPolygonDrawingOverlay(
-    GeofenceMapController controller,
-    int pointCount,
-  ) {
-    final theme = Theme.of(context);
-    final instructionText = pointCount >= 3
-        ? 'সংরক্ষণ করতে "সংরক্ষণ করুন" চাপুন।'
-        : 'কমপক্ষে ৩টি পয়েন্ট প্রয়োজন।';
-    return Positioned(
-      left: 16,
-      right: 16,
-      bottom: 16,
-      child: Material(
-        elevation: 10,
-        borderRadius: BorderRadius.circular(18),
-        color: theme.colorScheme.surface.withOpacity(0.95),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Text(
-                'নতুন পলিগন আঁকা হচ্ছে',
-                style: theme.textTheme.titleMedium?.copyWith(
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'মানচিত্রে ট্যাপ করে পয়েন্ট যোগ করুন। $instructionText',
-                style: theme.textTheme.bodyMedium,
-              ),
-              const SizedBox(height: 12),
-              Wrap(
-                spacing: 8,
-                runSpacing: 8,
-                alignment: WrapAlignment.end,
-                children: [
-                  OutlinedButton.icon(
-                    onPressed: pointCount > 0 ? _undoPolygonPoint : null,
-                    icon: const Icon(Icons.undo),
-                    label: const Text('আনডু'),
-                  ),
-                  OutlinedButton.icon(
-                    onPressed: _cancelPolygonDrawing,
-                    icon: const Icon(Icons.close),
-                    label: const Text('বাতিল'),
-                  ),
-                  FilledButton.icon(
-                    onPressed: pointCount >= 3
-                        ? () => _finishPolygonDrawing(controller)
-                        : null,
-                    icon: const Icon(Icons.check),
-                    label: const Text('সংরক্ষণ করুন'),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 
@@ -1461,56 +1163,6 @@ class _CollapsibleInformationButton extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class _PolygonColorPreview extends StatelessWidget {
-  const _PolygonColorPreview({
-    required this.color,
-    required this.label,
-    required this.selected,
-  });
-
-  final Color color;
-  final String label;
-  final bool selected;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final double size = selected ? 48 : 44;
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        AnimatedContainer(
-          duration: const Duration(milliseconds: 180),
-          width: size,
-          height: size,
-          decoration: BoxDecoration(
-            color: color,
-            shape: BoxShape.circle,
-            border: Border.all(
-              color: selected
-                  ? theme.colorScheme.onSurface.withOpacity(0.9)
-                  : Colors.white,
-              width: selected ? 3 : 2,
-            ),
-            boxShadow: [
-              BoxShadow(
-                color: color.withOpacity(selected ? 0.45 : 0.3),
-                blurRadius: selected ? 10 : 6,
-                offset: const Offset(0, 3),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(height: 6),
-        Text(
-          label,
-          style: theme.textTheme.labelSmall,
-        ),
-      ],
     );
   }
 }
